move talks data to data folder
